refactor: models and apis 폴더생성 및 컴포넌트 분리
