--- src/Models.ts
@@ -1,31 +1,0 @@
-export interface Order {
-  id: number;
-  orderAt: Date;
-  amount: number;
-  products: OrderProduct[];
-  shipping: Shipping;
-}
-
-export interface OrderProduct {
-  id: number
-  name: string
-  price: number
-  imageUrls: string[]
-  stock: Stock
-}
-
-export interface Stock {
-  color: string
-  band: string
-  cup: string
-  quantity: number
-}
-
-export interface Shipping {
-  id: number,
-  trackingNumber: string,
-  shippingFee: number,
-  address: string,
-  post: string,
-  message: string
-}
