+++ src/models/Order.ts
@@ -1,0 +1,10 @@
+import { OrderProduct } from "./Product";
+import { Shipping } from "./Shipping";
+
+export interface Order {
+  id: number;
+  orderAt: Date;
+  amount: number;
+  products: OrderProduct[];
+  shipping: Shipping;
+}
