+++ src/models/Product.ts
@@ -1,0 +1,14 @@
+export interface Stock {
+  color: string;
+  band: string;
+  cup: string;
+  quantity: number;
+}
+
+export interface OrderProduct {
+  id: number
+  name: string
+  price: number
+  imageUrls: string[]
+  stock: Stock
+}
