+++ src/models/Shipping.ts
@@ -1,0 +1,8 @@
+export interface Shipping {
+  id: number,
+  trackingNumber: string,
+  shippingFee: number,
+  address: string,
+  post: string,
+  message: string
+}
